requirements.txt
Werkzeug==2.2.2
Flask==2.2.2
authlib==1.2.0
requests-oauthlib==1.3.0
psycopg2-binary==2.9.9
pyjwt==2.7.0
google-generativeai
